refactor: extract stringToPath into a separate helper
